Keeps both sides: upstream's settings rework, model_render rework, mesh-tag change-gating,
static_gx node move + diagnostics span, WorldPointLight, brightness/weather options,
weapon trails; and the MONKEY lighting stack (sun shadow lanes, dynamic interiors, cube-map
torch shadows, daylight fixtures, spell lights, bake floor, K=12 exterior selection, debug panel).
All MONKEY light lanes migrated from Bevy PointLight to upstream's WorldPointLight.
1091 world+formats tests green; all 8 WGSL probes OK; server-less smoke (inn, Northshire) clean.

+++ crates/benilla-protocol/src/messages/group.rs
@@ -295,12 +295,14 @@ impl PartyMemberStatsInfo {
     /// [`power_display_scale`], exactly as the live-object leg divides. Without it an out-of-range
     /// warrior's rage bar reads ten times an in-range one's. Miss ⇒ `0`, the binding's own.
     pub fn shown_power(&self) -> u32 {
-        u32::from(self.cur_power.unwrap_or(0)) / power_display_scale(self.shown_power_type())
+        u32::from(self.cur_power.unwrap_or(0))
+            / power_display_scale(u32::from(self.shown_power_type()))
     }
 
     /// `UnitManaMax` on the record path (`0x5178af`), the same divide.
     pub fn shown_max_power(&self) -> u32 {
-        u32::from(self.max_power.unwrap_or(0)) / power_display_scale(self.shown_power_type())
+        u32::from(self.max_power.unwrap_or(0))
+            / power_display_scale(u32::from(self.shown_power_type()))
     }
 }
